+++ libs/domain/search/box/box.component.spec.ts
