+++ libs/domain/search/box/box.component.ts
