+++ libs/domain/search/box/box.model.ts
