+++ libs/domain/search/box/box.schema.ts
