+++ libs/domain/search/box/controllers/query-controls.controller.ts
