+++ libs/domain/search/box/controllers/render-suggestion.controller.ts
@@ -1,11 +1,11 @@
 import { Product, ProductMediaContainerSize } from '@spryker-oryx/product';
+import { Suggestion, SuggestionResource } from '@spryker-oryx/search';
 import { SemanticLinkType } from '@spryker-oryx/site';
 import { IconTypes } from '@spryker-oryx/ui/icon';
 import { i18n } from '@spryker-oryx/utilities';
 import { LitElement, TemplateResult } from 'lit';
 import { DirectiveResult } from 'lit/directive';
 import { html } from 'lit/static-html.js';
-import { Suggestion, SuggestionResource } from '../../../src';
 import { SearchBoxProperties } from '../box.model';
 
 interface LinksSection {
