+++ libs/domain/search/box/controllers/scrolling-area.controller.spec.ts
