+++ libs/domain/search/box/controllers/scrolling-area.controller.ts
