+++ libs/domain/search/box/index.ts
@@ -1,1 +1,4 @@
-export * from './src';
+export * from './box.component';
+export * from './box.model';
+export * from './controllers';
+export * from './styles';
