--- libs/domain/search/box/src/index.ts
@@ -1,4 +1,0 @@
-export * from './box.component';
-export * from './box.model';
-export * from './controllers';
-export * from './styles';
