+++ libs/domain/search/box/stories/box.stories.ts
