+++ libs/domain/search/box/styles/base.styles.ts
