+++ libs/domain/search/box/styles/index.ts
