+++ libs/domain/search/box/styles/oryx.styles.ts
