+++ libs/domain/search/src/components.ts
@@ -1,4 +1,4 @@
-export * from '../box/src/box.def';
+export * from '../box/box.def';
 export * from '../facet-color/facet-color.def';
 export * from '../facet-navigation/facet-navigation.def';
 export * from '../facet-value-navigation/facet-value-navigation.def';
